rename user file to auth controller file
